refactor: move settings-related logic to a new `Settings` module
- Extracted `Settings` types and utility functions from `Brew` to a new `Settings` module for better separation of concerns.
- Updated imports across the project to reflect the new structure.
- Added a `previewCups` state in the `Settings` page to allow dynamic updates for cup previews.
